+++ src/navigators/SignUpNavigator.js
@@ -1,21 +1,21 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import SignUpScreenFirst from '../screens/SignUp/SignUpScreenFirst';
-import SignUpScreenSecond from '../screens/SignUp/SignUpScreenSecond';
-import SignUpScreenThird from '../screens/SignUp/SignUpScreenThird';
-import SignUpScreenFourth from '../screens/SignUp/SignUpScreenFourth';
-import SignUpScreenFifth from '../screens/SignUp/SignUpScreenFifth';
+import SignUpScreenSelect from '../screens/SignUp/SignUpScreenSelect';
+import SignUpScreenFillInfo from '../screens/SignUp/SignUpScreenFillInfo';
+// import SignUpScreenOptional from '../screens/SignUp/SignUpScreenOptional';
+import SignUpScreenComplete from '../screens/SignUp/SignUpScreenComplete';
+import SignUpScreenAgreement from '../screens/SignUp/SignUpScreenAgreement';
 
 const Stack = createNativeStackNavigator();
 
 export default SignUpNavigator = () => {
   return (
     <Stack.Navigator>
-      <Stack.Screen name='SignUpScreenFirst' component={SignUpScreenFirst} options={{ headerShown: false}}/>
-      <Stack.Screen name='SignUpScreenSecond' component={SignUpScreenSecond} options={{ headerShown: false}}/>
-      <Stack.Screen name='SignUpScreenThird' component={SignUpScreenThird} options={{ headerShown: false}}/>
-      <Stack.Screen name='SignUpScreenFourth' component={SignUpScreenFourth} options={{ headerShown: false}}/>
-      <Stack.Screen name='SignUpScreenFifth' component={SignUpScreenFifth} options={{ headerShown: false}}/>
+      <Stack.Screen name='SignUpScreenSelect' component={SignUpScreenSelect} options={{ headerShown: false}}/>
+      <Stack.Screen name='SignUpScreenAgreement' component={SignUpScreenAgreement} options={{ headerShown: false}}/>
+      <Stack.Screen name='SignUpScreenFillInfo' component={SignUpScreenFillInfo} options={{ headerShown: false}}/>
+      <Stack.Screen name='SignUpScreenComplete' component={SignUpScreenComplete} options={{ headerShown: false}}/>
+      {/* <Stack.Screen name='SignUpScreenOptional' component={SignUpScreenOptional} options={{ headerShown: false}}/> */}
     </Stack.Navigator>
   )
 }
